--- utbot-python/src/main/kotlin/org/utbot/python/fuzzing/value/UndefValue.kt
@@ -1,6 +1,0 @@
-package org.utbot.python.fuzzing.value
-
-import org.utbot.fuzzing.Mutation
-import org.utbot.fuzzing.seeds.KnownValue
-
-class ObjectValue : KnownValue<ObjectValue>
