+++ utbot-python/src/main/kotlin/org/utbot/python/newtyping/mypy/MypyAnnotations.kt
@@ -250,3 +250,14 @@ val annotationAdapter: PolymorphicJsonAdapterFactory<MypyAnnotationNode> =
         .withSubtype(PythonNoneType::class.java, AnnotationType.NoneType.name)
         .withSubtype(TypeAliasNode::class.java, AnnotationType.TypeAlias.name)
         .withSubtype(UnknownAnnotationNode::class.java, AnnotationType.Unknown.name)
+
+object MypyAnnotations {
+
+    data class MypyReportLine(
+        val line: Int,
+        val type: String,
+        val message: String,
+        val file: String
+    )
+
+}
